Mejora de velocidades

+++ P3/objetos_B3.h
@@ -328,7 +328,7 @@ class _camionbomberos: public _triangulos3D{
 
        float giro_tubo;
        int giro_ruedas;
-       int giro_plataforma = 0.0;
+       float giro_plataforma = 0.0;
        float translacion_escalera = 0.0;
        float translacion_escalera_min = 0.0;
        float translacion_escalera_max = 2.3;
@@ -360,3 +360,4 @@ class _camionbomberos: public _triangulos3D{
        _pistola pistola;
        _agua agua;
 };
+
